build: add junit

+++ src/main/java/com/ricka/princy/bonjouraurevoir/endpoint/rest/security/readme/monitor/model/entry/ReadmeEntryResponseContent.java
@@ -10,7 +10,7 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @Builder
 @JsonInclude(NON_NULL)
 public record ReadmeEntryResponseContent(
-    BigInteger size,
+    Integer size,
     String mimeType,
     String text,
     String encoding
